+++ missile-launcher-bot/remote.config.js
@@ -1,0 +1,11 @@
+const _TARGET_IP = '172.20.0.20';
+const _TARGET_PORT = '80';
+const _COMMMANDS_ENDPOINT = '/command'
+
+export const SERVER_CONFIG = {
+    TARGET_IP : _TARGET_IP,
+    TARGET_PORT: _TARGET_PORT,
+    ENDPOINTS: {
+        COMMMAND: _COMMMANDS_ENDPOINT
+    }
+}
