+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "frostburn-missile-launcher",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "Start frostburn via Discord",
   "main": "./missile-launcher-bot/main.js",
   "scripts": {
